move Alert class and related to independent module
